+++ plugins/plugin-iter8/src/models/exprSetup.ts
@@ -1,10 +1,6 @@
 import { ModeRegistration } from '@kui-shell/core'
 import { KubeResource } from '@kui-shell/plugin-kubectl'
-import { renderForm } from "./exprForm"
-// Interface used by Custom Kubernetes Resource (Experiment)
-interface ExprResource extends KubeResource {
-  status: any;
-}
+import { renderForm } from '../modes/exprForm'
 
 /*
 * Checks the correct command response 
@@ -16,9 +12,9 @@ function verifyResponse(resource: KubeResource): boolean {
 /*
 *	Renders the Form for the Sidecar
 */
-async function renderView(){
-	return renderForm();
-}
+// async function renderView(){
+// 	return renderForm();
+// }
 /*
  * This is the model for the View rendering
  */
@@ -26,7 +22,7 @@ const exprcreateMode: ModeRegistration<KubeResource> = {
 	when: verifyResponse,
 	mode: {
 		mode: "Experiment Setup",
-		content: renderView,
+		content: renderForm,
 	}
 }
