+++ plugins/plugin-iter8/src/web/scss/static/exprForm.scss
@@ -1,11 +1,5 @@
 
 //Defined colors
-$dark-brown: #17252A;
-$dark-green: #2B7A78;
-$mint-green: #3AAFA9;
-$light-green: #DEF2F1;
-$light-red: #ff6961;
-$dark-red: #ff1f13;
 $light-grey: #c6ccd8;
 $trans-grey: #5E6367;
